Loading several action bars no longer freezes client
This was occurring due to the interactions between the various Pickup functions and the action bars. While it was hardly noticeable for a single bar, every additional shared bar beyond 1 created very noticeable lag for several seconds or more on slower computers with even some high end computers freezing for a few seconds during processing.

Offloading the work to a coroutine on a timer isn't ideal but it dramatically lessens load times and client stuttering.