Should set base spell if actionslot ID isn't found
some abilities have versions that override some base spell. Picking up the spell from the spellbook would provide one ID, but calling GetActionInfo() for the slot would return a completely different ID. For example, Druid ability "Swipe" is in the spellbook as spellID 213764 but returns 106785 when you call GetActionInfo() for the slot.