Added optional minimap icon
fixes #9. The right-click context menu is only set up to allow on-demand reloading of bars, but it can be extended later with more actions. Left-clicking the button will open the addon settings window.

This implementation relies on LibDBIcon so it should also be compatible with various frame handlers like Titan Panel.